test(hermes-plugin): colocate Stream 4 tool tests in test_tools.py
Move the 34 lifecycle + template tests from
`test_tools_lifecycle_templates.py` into the canonical `test_tools.py`
so DoD review finds them in the same file as the Stream 1/3/5 tests.

No behavior change — identical test cases, just relocated. All
34 tests run and pass against the dedicated worktree venv. The sibling
file is deleted.

Covers AC-027..AC-035, AC-051..AC-053, AC-073..AC-077 exactly as in
commit 66b69cd.
